Fix isNew for documents.
We don’t want to specify that idAttribute is _id, because then _id would have to be compatible with Backbone ids, and would have to work in URLs, which isn’t the case with (most) Mongo IDs. So instead, we fake it by having an id property and an _id attribute which aren’t necessarily linked.